refactor(scripts): extract Discord REST client into lib/discord-api.ts
Move-only extraction of the Discord API client layer (rate-limited
fetch with 429 retry, pagination over members/threads/messages,
permission/overwrite resolution, and the request/response types)
from telemetry/capture-discord-server-metrics.ts into
lib/discord-api.ts. The capture script keeps CLI parsing, count
merging/aggregation, and history persistence orchestration.

The only non-move change is an explicit `string | null` annotation
on lastMessageId in fetchChannelMessages, fixing the pre-existing
TS7022 narrowing-circularity error; the annotation equals the
declared type of its initializer, so behavior is unchanged.
